redis: optional redis auth env var and secret
intent to the consumers of your library:
correlates with PATCH in semantic versioning).
codebase (this correlates with MINOR in semantic versioning).
beginning of its optional body or footer section introduces a breaking 
API change (correlating with MAJOR in semantic versioning). A BREAKING 
CHANGE can be part of commits of any type.
example @commitlint/config-conventional (based on the the Angular 
convention) recommends chore:, docs:, style:, refactor:, perf:, test:, 
and others.